+++ src/main/java/tw/kid7/BannerMaker/command/HandCommand.java
@@ -1,0 +1,27 @@
+package tw.kid7.BannerMaker.command;
+
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandSender;
+
+class HandCommand extends AbstractCommand {
+    //名稱
+    private static final String name = "Hand";
+    //介紹
+    private static final String description = "Show banner info of the banner in hand";
+    //權限
+    private static final String permission = "BannerMaker.hand";
+    //使用方法
+    private static final String usage = "/bm hand";
+    //僅能由玩家執行
+    private static final boolean onlyFromPlayer = true;
+
+    HandCommand() {
+        super(name, description, permission, usage, onlyFromPlayer);
+    }
+
+    @Override
+    boolean handle(CommandSender sender, Command command, String label, String[] args) {
+        //TODO
+        return true;
+    }
+}